+++ packages/xrpl/src/models/ledger/LedgerEntry.ts
@@ -1,6 +1,7 @@
 import AccountRoot from './AccountRoot'
 import Amendments from './Amendments'
 import AMM from './AMM'
+import Bridge from './Bridge'
 import Check from './Check'
 import DepositPreauth from './DepositPreauth'
 import DirectoryNode from './DirectoryNode'
@@ -13,11 +14,14 @@ import PayChannel from './PayChannel'
 import RippleState from './RippleState'
 import SignerList from './SignerList'
 import Ticket from './Ticket'
+import XChainOwnedClaimID from './XChainOwnedClaimID'
+import XChainOwnedCreateAccountClaimID from './XChainOwnedCreateAccountClaimID'
 
 type LedgerEntry =
   | AccountRoot
   | Amendments
   | AMM
+  | Bridge
   | Check
   | DepositPreauth
   | DirectoryNode
@@ -30,5 +34,7 @@ type LedgerEntry =
   | RippleState
   | SignerList
   | Ticket
+  | XChainOwnedClaimID
+  | XChainOwnedCreateAccountClaimID
 
 export default LedgerEntry