+++ packages/xrpl/src/models/transactions/offerCreate.ts
@@ -1,4 +1,3 @@
-/* eslint-disable complexity -- Necessary for validateOfferCreate */
 import { ValidationError } from '../../errors'
 import { Amount } from '../common'