+++ packages/xrpl/HISTORY.md
@@ -1,13 +1,21 @@
 # xrpl.js (ripple-lib) Release History
 
 Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xrpl-announce) for release announcements. We recommend that xrpl.js (ripple-lib) users stay up-to-date with the latest stable release.
+
 ## Unreleased
 
 ### Added
 * Add support for Concise Transaction Identifier (ctid) as defined in [XLS-37](https://github.com/XRPLF/XRPL-Standards/discussions/91)
 
-## Fixed
+## 2.12.0 (2023-09-27)
+### Added
+* Added `ports` field to `ServerInfoResponse`
+* Support for the XChainBridge amendment.
+
+### Fixed
 * Fix request model fields related to AMM
+* Rename `AMMAccount` to `Account` on `AMM` ledger objects
+* Fixed `EscrowCancel` and `EscrowFinish` validation
 
 ## 2.11.0 (2023-08-24)